# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ADEX documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "ADEX API Server"
        }
    ],
    "paths": {
...
        "/chat": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Создание чата исполнителем, когда оставляет отклик",
                "operationId": "createChat",
                "requestBody": {
                    "description": "Заполните поля для создания чата",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "task_id",
                                    "text"
                                ],
                                "properties": {
                                    "сustomer_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "task_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Это отклик!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unautorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unautorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/chat/{chatId}/message": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Добавление сообщения в чат",
                "operationId": "createMessage",
                "requestBody": {
                    "description": "Заполните поля для добавления сообщения в чат",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Текст сообщения"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "Это текст сообщения!"
                                        },
                                        "author": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "userreg1"
                                                },
                                                "firstname": {
                                                    "type": "string",
                                                    "example": "Ivan"
                                                },
                                                "lastname": {
                                                    "type": "string",
                                                    "example": "Ivanovich"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "chat_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-04-04"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unautorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unautorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}